backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
firebase-admin==6.6.0
google-cloud-firestore==2.16.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
markdown-it-py==3.0.0
bleach==6.1.0
black==24.10.0
ruff==0.6.9
